move server/schedule/storelimit to server/core/storelimit
