+++ server/schedule/operator/influence.go
@@ -15,7 +15,7 @@ package operator
 
 import (
 	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/schedule/storelimit"
+	"github.com/tikv/pd/server/core/storelimit"
 )
 
 // OpInfluence records the influence of the cluster.
